Eliminación de los requisitos de seguridad no esenciales
Se dejan de utilizar JWT, authentication manager y password encoder. Si sobrara tiempo se revisaría para añadirlos de nuevo.

+++ src/main/java/com/unizar/unozar/core/service/PlayerServiceImpl.java
@@ -1,74 +1,47 @@
 package com.unizar.unozar.core.service;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 
-import javax.servlet.http.HttpServletRequest;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.unizar.unozar.core.CustomPlayerDetailsService;
-import com.unizar.unozar.core.JWTUtil;
 import com.unizar.unozar.core.DTO.PlayerDTO;
 import com.unizar.unozar.core.controller.resources.AuthenticationRequest;
 import com.unizar.unozar.core.controller.resources.AuthenticationResponse;
-import com.unizar.unozar.core.controller.resources.BasicPlayerRequest;
+import com.unizar.unozar.core.controller.resources.CreatePlayerRequest;
 import com.unizar.unozar.core.controller.resources.DeletePlayerRequest;
+import com.unizar.unozar.core.controller.resources.RefreshTokenRequest;
+import com.unizar.unozar.core.controller.resources.UpdatePlayerRequest;
 import com.unizar.unozar.core.entities.Player;
-import com.unizar.unozar.core.exceptions.AliasSyntaxError;
 import com.unizar.unozar.core.exceptions.EmailInUse;
-import com.unizar.unozar.core.exceptions.EmailSyntaxError;
-import com.unizar.unozar.core.exceptions.EmptyRequest;
+import com.unizar.unozar.core.exceptions.InvalidIdentity;
 import com.unizar.unozar.core.exceptions.InvalidPassword;
-import com.unizar.unozar.core.exceptions.PasswordSyntaxError;
 import com.unizar.unozar.core.exceptions.PlayerNotFound;
 import com.unizar.unozar.core.repository.PlayerRepository;
-
-import io.jsonwebtoken.impl.DefaultClaims;
 
 @Service
 public class PlayerServiceImpl implements PlayerService{
 
   @Autowired
   private PlayerRepository playerRepository;
-  
-  @Autowired
-  private JWTUtil jWTUtil;
-  
-  @Autowired
-  private PasswordEncoder bCryptEncoder;
-  
-  @Autowired
-  private CustomPlayerDetailsService playerDetailsService;
-  
-  @Autowired
-  private AuthenticationManager authenticationManager;
-  
+
   @Override
-  public PlayerDTO createPlayer(BasicPlayerRequest request){
-    checkCreatePlayerRequest(request);
+  public PlayerDTO createPlayer(CreatePlayerRequest request){
     String emailNewUser = request.getEmail();
     Optional<Player> toFind = playerRepository.findByEmail(emailNewUser);
     if(toFind.isPresent()){
       throw new EmailInUse("The email is already in use");
     }
     Player created = playerRepository.save(new Player(request.getEmail(), 
-        request.getAlias(), bCryptEncoder.encode(request.getPassword())));
+        request.getAlias(), request.getPassword()));
     PlayerDTO player = new PlayerDTO(created);
     return player;
   }
   
   @Override
   public PlayerDTO readPlayer(String id){
-    Optional<Player> toFind = playerRepository.findByEmail(id);
-    if(!toFind.isPresent()){
+    Optional<Player> toFind = playerRepository.findById(id);
+    if(toFind.isEmpty()){
       throw new PlayerNotFound("Id does not exist in the system");
     }
     Player toRead = toFind.get();
@@ -76,13 +49,19 @@ public class PlayerServiceImpl implements PlayerService{
   }
 
   @Override
-  public Void updatePlayer(String id, BasicPlayerRequest request){
-    checkUpdatePlayerRequest(request);
+  public Void updatePlayer(String id, UpdatePlayerRequest request){
+    if(id != request.getToken().substring(0, 35)){
+      throw new InvalidIdentity("The requester's id does not match with the " +
+          "given id");      
+    }    
     Optional<Player> toFind = playerRepository.findById(id);
     if (toFind.isPresent()){
       throw new PlayerNotFound("Id does not exist in the system");
     }
     Player toUpdate = toFind.get();
+    if(!toUpdate.checkSession(request.getToken().substring(36))){
+      throw new InvalidIdentity("Invalid token");
+    }
     if(request.getAlias() != null){
       toUpdate.setAlias(request.getAlias());
     }
@@ -95,13 +74,25 @@ public class PlayerServiceImpl implements PlayerService{
       toUpdate.setEmail(request.getEmail());
     }
     if(request.getPassword() != null){
-      toUpdate.setPassword(bCryptEncoder.encode(request.getPassword()));
+      toUpdate.setPassword(request.getPassword());
     }
     return null;
   }
 
   @Override
   public Void deletePlayer(String id, DeletePlayerRequest request){
+    if(id != request.getToken().substring(0, 35)){
+      throw new InvalidIdentity("The requester's id does not match with the " +
+          "given id");      
+    }   
+    Optional<Player> toFind = playerRepository.findById(id);
+    if (toFind.isPresent()){
+      throw new PlayerNotFound("Id does not exist in the system");
+    }
+    Player toDelete = toFind.get();
+    if(!toDelete.checkSession(request.getToken().substring(36))){
+      throw new InvalidIdentity("Invalid token");
+    }
     playerRepository.deleteById(id);
     return null;
   }
@@ -114,84 +105,27 @@ public class PlayerServiceImpl implements PlayerService{
       throw new PlayerNotFound("Email does not exist in the system");
     }
     Player toAuth = toFind.get();
-    if(!toAuth.isValidPassword(bCryptEncoder.encode(request.getPassword()))){
+    if(!toAuth.isValidPassword(request.getPassword())){
       throw new InvalidPassword("Invalid password");
     }
-    authenticationManager.authenticate(
-        new UsernamePasswordAuthenticationToken(
-            toAuth.getId(), toAuth.getPassword()));
-    UserDetails detailsFromToAuth = 
-        playerDetailsService.loadUserByUsername(toAuth.getId());
-    String token = jWTUtil.generateToken(detailsFromToAuth);
+    String token = toAuth.getId() + toAuth.updateSession();
     return new AuthenticationResponse(token);
   }
   
   @Override
   public AuthenticationResponse 
-      refreshToken(HttpServletRequest request){
-    DefaultClaims claims = (io.jsonwebtoken.impl.DefaultClaims) 
-        request.getAttribute("claims");
-    Map<String, Object> expectedMap = getMapFromTokenClaims(claims);
-    String token = jWTUtil.doGenerateRefreshToken(expectedMap, 
-        (String) expectedMap.get("sub"));
+      refreshToken(RefreshTokenRequest request){
+    String id = request.getToken().substring(0, 35);
+    Optional<Player> toFind = playerRepository.findById(id);
+    if (toFind.isPresent()){
+      throw new PlayerNotFound("Id does not exist in the system");
+    }
+    Player toRefresh = toFind.get();
+    if(!toRefresh.checkSession(request.getToken().substring(36))){
+      throw new InvalidIdentity("Invalid token");
+    }
+    String token = toRefresh.getId() + toRefresh.updateSession();
     return new AuthenticationResponse(token);
-  }
-  
-  private Void checkCreatePlayerRequest(BasicPlayerRequest request){
-    checkEmail(request.getEmail());
-    checkAlias(request.getAlias());
-    checkPassword(request.getPassword());
-    return null;
-  }
-
-  private Void checkUpdatePlayerRequest(BasicPlayerRequest request) {
-    boolean requestEmpty = true;
-    if(request.getEmail() != null){
-      requestEmpty = false;
-      checkEmail(request.getEmail());
-    }
-    if(request.getAlias() != null){
-      requestEmpty = false;
-      checkAlias(request.getAlias());
-    }
-    if(request.getPassword() != null){
-      requestEmpty = false;
-      checkPassword(request.getPassword());
-    }
-    if(requestEmpty){
-      throw new EmptyRequest("The update player request is empty");
-    }
-    return null;
-  }
-
-  
-  private Void checkPassword(String password){
-    if(password.length() <= 3){
-      throw new PasswordSyntaxError("The password is too short");
-    }
-    return null;
-  }
-
-  private Void checkAlias(String alias){
-    if(alias.length() <= 3){
-      throw new AliasSyntaxError("The alias is too short");
-    }
-    return null;
-  }
-
-  private Void checkEmail(String email){
-    if(!email.contains("@")){
-      throw new EmailSyntaxError("The email is not a valid email direction");
-    }
-    return null;
-  }
-
-  private Map<String, Object> getMapFromTokenClaims(DefaultClaims claims){
-    Map<String, Object> expectedMap = new HashMap<String, Object>();
-    for(Entry<String, Object> entry : claims.entrySet()){
-      expectedMap.put(entry.getKey(), entry.getValue());
-    }
-    return expectedMap;
   }
   
 }
